Setup unit testing with jest server
