--- server/dummy.js
@@ -1,5 +1,0 @@
-function sayHello(name) {
-    return `Hi, ${name}.`
-}
-
-module.exports = sayHello;
