+++ server/package.json
@@ -6,7 +6,21 @@
   "scripts": {
     "start": "node index",
     "server": "nodemon index",
-    "test": "jest --coverage"
+    "test": "jest --detectOpenHandles --forceExit"
+  },
+  "jest": {
+    "collectCoverage": true,
+    "testRegex": "((\\.|/*.)(test))\\.js?$",
+    "collectCoverageFrom": [
+      "routes/*.{js,jsx}",
+      "models/*.{js,jsx}",
+      "middleware/*.{js,jsx}",
+      "database/*.{js,jsx}"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/jspm_packages"
+    ]
   },
   "author": "Asad Mahmood, Stefan Radovic and Dennis Kurtz",
   "license": "ISC",
@@ -19,12 +33,7 @@
     "mysql2": "^2.2.5",
     "nodemon": "^2.0.7",
     "sequelize": "^6.5.0",
-    "sequelize-cli": "^6.2.0"
-  },
-  "jest": {
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "/jspm_packages"
-    ]
+    "sequelize-cli": "^6.2.0",
+    "supertest": "^6.1.3"
   }
 }
