--- server/tests/dummy.test.js
@@ -1,5 +1,0 @@
-const sayHello = require('../dummy');
-
-test('returns a personalized greeting', () => {
-  expect(sayHello('Asad')).toBe("Hi, Asad.");
-});
